requirements.txt
Django==1.3.1
django-mediagenerator==1.11
django-tastypie==0.9.11
ipdb==0.6.1
ipython==0.12
mimeparse==0.1.3
python-dateutil==1.5
readline==6.2.1
wsgiref==0.1.2
